Refactored to move files out of tests
Created persistence and handlers libs to organize code better
