+++ pethome.sln
@@ -7,6 +7,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "view", "src\view\view.cspro
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tests", "src\tests\tests.csproj", "{CFF79C91-6A5D-44F9-9C8E-C9AB94DCF7CF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "persistence", "src\persistence\persistence.csproj", "{ECFB2C4D-9A00-4E9D-B5FE-0D2A76D50AE1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "handlers", "src\handlers\handlers.csproj", "{6CD1D42A-B830-47C9-82B7-2B8173F587ED}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +25,14 @@ Global
 		{CFF79C91-6A5D-44F9-9C8E-C9AB94DCF7CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CFF79C91-6A5D-44F9-9C8E-C9AB94DCF7CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CFF79C91-6A5D-44F9-9C8E-C9AB94DCF7CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ECFB2C4D-9A00-4E9D-B5FE-0D2A76D50AE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ECFB2C4D-9A00-4E9D-B5FE-0D2A76D50AE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ECFB2C4D-9A00-4E9D-B5FE-0D2A76D50AE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ECFB2C4D-9A00-4E9D-B5FE-0D2A76D50AE1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6CD1D42A-B830-47C9-82B7-2B8173F587ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CD1D42A-B830-47C9-82B7-2B8173F587ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CD1D42A-B830-47C9-82B7-2B8173F587ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CD1D42A-B830-47C9-82B7-2B8173F587ED}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
