+++ src/persistence/Models/Animal.cs
@@ -1,8 +1,11 @@
-namespace tests
+using System;
+
+namespace PetHome.Persistence.Models
 {
-    public class RegisterNewAnimal
+    public class Animal
     {
         public string Species { get; set; }
+        public DateTime Entered { get; set; }
         public string Color { get; set; }
         public string Breed { get; set; }
         public string Gender { get; set; }
