+++ src/persistence/persistence.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <AssemblyName>PetHome.Persistence</AssemblyName>
+    <RootNamespace>PetHome.Persistence</RootNamespace>
+  </PropertyGroup>
+
+</Project>
