+++ src/tests/tests.csproj
@@ -14,4 +14,9 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\handlers\handlers.csproj" />
+    <ProjectReference Include="..\persistence\persistence.csproj" />
+  </ItemGroup>
+
 </Project>
